Refactor and reformat user and scripts to allow workspace for team
